nec plugin: allow to delete resource with ERROR status
Previously if a resource is in ERROR status and there is no
corresponding resource on OpenFlow controller, the resource
cannot be deleted through an API request.
This commit rearrange ERROR status check to allow resource
with ERROR status to be deleted.

Closes-Bug: #1295754

+++ neutron/plugins/nec/router_drivers.py
@@ -162,6 +162,8 @@ class RouterOpenFlowDriver(RouterDriverBase):
 
     @call_log.log
     def delete_router(self, context, router_id, router):
+        if not self.ofc.exists_ofc_router(context, router_id):
+            return
         try:
             self.ofc.delete_ofc_router(context, router_id, router)
         except (nexc.OFCException, nexc.OFCMappingNotFound) as exc:
